server/requirements.txt
flask==3.1.0
torch==2.2.2
torchvision==0.17.2
Pillow==10.3.0
requests==2.32.3
pandas==2.2.2

firebase-admin==6.4.0
google-cloud-vision==3.10.1
google-auth==2.29.0
google-api-core==2.24.0

openai==1.30.1
